CashFlows/requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
Django==3.0.2
django-jsonfield==1.4.0
fuzzywuzzy==0.17.0
idna==2.8
numpy==1.18.1
pandas==1.0.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.8
xlrd==1.2.0
